+++ sprockets/mixins/metrics/__init__.py
@@ -1,2 +1,5 @@
+from .statsd import StatsdMixin
+
 version_info = (0, 0, 0)
 __version__ = '.'.join(str(v) for v in version_info)
+__all__ = ['StatsdMixin']