+++ docs/history.rst
@@ -5,5 +5,7 @@ Release History
 
 `Next Release`_
 ---------------
+- Add :class:`sprockets.mixins.metrics.StatsdMixin`
+- Add :class:`sprockets.mixins.metrics.testing.FakeStatsdServer`
 
 .. _Next Release: https://github.com/sprockets/sprockets.mixins.metrics/compare/0.0.0...master